- running rand_gen and rand_access tests - rand gen type fix

+++ benchmarks_new/rand_generator/rand_generator_main.h
@@ -7,7 +7,7 @@
 #include <chrono>
 
 
-typedef float base_type;
+typedef double base_type;
 
 #include "rand_generator.h"
